fixed: scan qr code don't add already logged in account
Closes #519
