+++ src/ui/views/Welcome/Sync/SyncQr.tsx
@@ -1,51 +1,13 @@
 import { Typography, CardMedia } from '@mui/material';
 import { makeStyles } from '@mui/styles';
 import { Box } from '@mui/system';
-import { Core } from '@walletconnect/core';
-import SignClient from '@walletconnect/sign-client';
-import { type SessionTypes } from '@walletconnect/types';
-import * as bip39 from 'bip39';
-import HDWallet from 'ethereum-hdwallet';
-import React, { useEffect, useCallback, useState } from 'react';
+import React from 'react';
 import { QRCode } from 'react-qrcode-logo';
 
-import { FCLWalletConnectMethod } from '@/shared/utils/type';
 import scanIcon from 'ui/assets/scan.svg';
 import lilo from 'ui/FRWAssets/image/lilo.png';
-import { useWallet } from 'ui/utils';
 
-interface AccountKey {
-  hashAlgo: number;
-  publicKey: string;
-  signAlgo: number;
-  weight: number;
-}
-
-interface DeviceInfoRequest {
-  deviceId: string;
-  ip: string;
-  name: string;
-  type: string;
-  userAgent: string;
-
-  continent?: string;
-  continentCode?: string;
-  country?: string;
-  countryCode?: string;
-  regionName?: string;
-  city?: string;
-  district?: string;
-  zip?: string;
-  lat?: number;
-  lon?: number;
-  timezone?: string;
-  currency?: string;
-  isp?: string;
-  org?: string;
-  device_id?: string;
-}
-
-const useStyles = makeStyles((theme) => ({
+const useStyles = makeStyles(() => ({
   customInputLabel: {
     '& legend': {
       visibility: 'visible',
@@ -66,219 +28,14 @@ const useStyles = makeStyles((theme) => ({
   },
 }));
 
-const SyncQr = ({
-  handleSwitchTab,
-  savedUsername,
-  confirmMnemonic,
-  setUsername,
-  setAccountKey,
-  setDeviceInfo,
-}) => {
-  const usewallet = useWallet();
+interface SyncQrProps {
+  uri: string;
+  loadingString: string | null;
+  secondLine: string;
+}
+
+const SyncQr = ({ uri, loadingString, secondLine }: SyncQrProps) => {
   const classes = useStyles();
-  const [Uri, setUri] = useState('');
-  const [loading, setShowLoading] = useState<boolean>(false);
-  const [session, setSession] = useState<SessionTypes.Struct>();
-  const [mnemonic, setMnemonic] = useState(bip39.generateMnemonic());
-  const [currentNetwork, setNetwork] = useState('mainnet');
-
-  const loadNetwork = useCallback(async () => {
-    const currentNetwork = await usewallet.getNetwork();
-    setNetwork(currentNetwork);
-  }, [usewallet]);
-
-  useEffect(() => {
-    loadNetwork();
-  }, [loadNetwork]);
-
-  const onSessionConnected = useCallback(async (_session: SessionTypes.Struct) => {
-    console.log('_session ', _session);
-    setShowLoading(true);
-    setSession(_session);
-  }, []);
-
-  const _subscribeToEvents = useCallback(
-    async (_client: SignClient) => {
-      if (typeof _client === 'undefined') {
-        throw new Error('WalletConnect is not initialized');
-      }
-
-      _client.on('session_update', ({ topic, params }) => {
-        console.log('EVENT', 'session_update', { topic, params });
-        const { namespaces } = params;
-        const _session = _client.session.get(topic);
-        const updatedSession = { ..._session, namespaces };
-        onSessionConnected(updatedSession);
-      });
-      console.log('EVENT _client ', _client);
-    },
-    [onSessionConnected]
-  );
-  const getAccountKey = useCallback(() => {
-    const hdwallet = HDWallet.fromMnemonic(mnemonic);
-    const publicKey = hdwallet.derive("m/44'/539'/0'/0/0").getPublicKey().toString('hex');
-    const key: AccountKey = {
-      hashAlgo: 1,
-      signAlgo: 2,
-      weight: 1000,
-      publicKey: publicKey,
-    };
-    return key;
-  }, [mnemonic]);
-
-  const getDeviceInfo = useCallback(async (): Promise<DeviceInfoRequest> => {
-    const result = await usewallet.openapi.getLocation();
-    const installationId = await usewallet.openapi.getInstallationId();
-    // console.log('location ', userlocation);
-    const userlocation = result.data;
-    const deviceInfo: DeviceInfoRequest = {
-      city: userlocation.city,
-      continent: userlocation.country,
-      continentCode: userlocation.countryCode,
-      country: userlocation.country,
-      countryCode: userlocation.countryCode,
-      currency: userlocation.countryCode,
-      deviceId: installationId,
-      device_id: installationId,
-      district: '',
-      ip: userlocation.query,
-      isp: userlocation.as,
-      lat: userlocation.lat,
-      lon: userlocation.lon,
-      name: 'FRW Chrome Extension',
-      org: userlocation.org,
-      regionName: userlocation.regionName,
-      type: '2',
-      userAgent: 'Chrome',
-      zip: userlocation.zip,
-    };
-    return deviceInfo;
-  }, [usewallet]);
-
-  const sendRequest = useCallback(
-    async (wallet: SignClient, topic: string) => {
-      console.log(wallet);
-      wallet
-        .request({
-          topic: topic,
-          chainId: `flow:${currentNetwork}`,
-          request: {
-            method: FCLWalletConnectMethod.accountInfo,
-            params: [],
-          },
-        })
-        .then(async (result: any) => {
-          console.log('result ', result);
-          const jsonObject = JSON.parse(result);
-          console.log('jsonObject ', jsonObject);
-          if (jsonObject.method === FCLWalletConnectMethod.accountInfo) {
-            const accountKey: AccountKey = getAccountKey();
-            const deviceInfo: DeviceInfoRequest = await getDeviceInfo();
-            const ak = {
-              public_key: accountKey.publicKey,
-              hash_algo: accountKey.hashAlgo,
-              sign_algo: accountKey.signAlgo,
-              weight: accountKey.weight,
-            };
-            console.log('sent ->', accountKey);
-            confirmMnemonic(mnemonic);
-            setAccountKey(ak);
-            setDeviceInfo(deviceInfo);
-            wallet
-              .request({
-                topic: topic,
-                chainId: `flow:${currentNetwork}`,
-                request: {
-                  method: FCLWalletConnectMethod.addDeviceInfo,
-                  params: {
-                    method: '',
-                    data: {
-                      username: '',
-                      accountKey: accountKey,
-                      deviceInfo: deviceInfo,
-                    },
-                  },
-                },
-              })
-              .then(async (sent) => {
-                handleSwitchTab();
-                // usewallet.signInV3(mnemonic, ak, deviceInfo).then(async (result) => {
-
-                //   const userInfo = await usewallet.getUserInfo(true);
-                //   setUsername(userInfo.username);
-                //   handleSwitchTab();
-                // }).catch((error) => {
-                //   console.error('Error in sign in wallet request:', error);
-                // });
-              })
-              .catch((error) => {
-                console.error('Error in second wallet request:', error);
-              });
-          }
-        })
-        .catch((error) => {
-          console.error('Error in first wallet request:', error);
-        });
-    },
-    [
-      confirmMnemonic,
-      currentNetwork,
-      getAccountKey,
-      getDeviceInfo,
-      handleSwitchTab,
-      mnemonic,
-      setAccountKey,
-      setDeviceInfo,
-    ]
-  );
-
-  useEffect(() => {
-    const createWeb3Wallet = async () => {
-      try {
-        const wallet = await SignClient.init({
-          // @ts-ignore: Unreachable code error
-          core: new Core({
-            projectId: process.env.WC_PROJECTID,
-          }),
-          metadata: {
-            name: 'Flow Walllet',
-            description: 'Digital wallet created for everyone.',
-            url: 'https://fcw-link.lilico.app',
-            icons: ['https://fcw-link.lilico.app/logo.png'],
-          },
-        });
-        await _subscribeToEvents(wallet);
-
-        try {
-          const { uri, approval } = await wallet.connect({
-            requiredNamespaces: {
-              flow: {
-                methods: [FCLWalletConnectMethod.accountInfo, FCLWalletConnectMethod.addDeviceInfo],
-                chains: [`flow:${currentNetwork}`],
-                events: [],
-              },
-            },
-          });
-
-          if (uri) {
-            console.log('uri ', uri);
-            await setUri(uri);
-            const session = await approval();
-            await onSessionConnected(session);
-            console.log('session ', session);
-            sendRequest(wallet, session.topic);
-          }
-        } catch (e) {
-          console.error(e);
-        }
-      } catch (e) {
-        console.error(e);
-      }
-    };
-
-    createWeb3Wallet();
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, []);
 
   return (
     <>
@@ -321,8 +78,6 @@ const SyncQr = ({
               variant="body1"
               sx={{ color: 'primary.light', fontSize: '16px', margin: '24px 0 32px' }}
             >
-              {/* {chrome.i18n.getMessage('appDescription')} {' '} */}
-
               {chrome.i18n.getMessage('Open_your_Flow_Reference_on_Mobil')}
             </Typography>{' '}
             <CardMedia
@@ -338,8 +93,6 @@ const SyncQr = ({
             />
           </Box>
           <Typography variant="body1" sx={{ color: '#8C9BAB', pt: '12px', fontSize: '12px' }}>
-            {/* {chrome.i18n.getMessage('appDescription')} {' '} */}
-
             {chrome.i18n.getMessage(' Note_Your_recovery_phrase_will_not')}
           </Typography>
         </Box>
@@ -353,17 +106,7 @@ const SyncQr = ({
             width: '347px',
           }}
         >
-          {/* <Box>
-              <Typography sx={{
-                width: '347px',
-                overflow: 'hidden',
-                whiteSpace: 'nowrap',
-                textOverflow: 'ellipsis'
-              }}>{Uri}</Typography>
-              <button onClick={copyToClipboard}>Copy Uri</button>
-              {copySuccess && <Box>{copySuccess}</Box>}
-            </Box> */}
-          {Uri && (
+          {uri && (
             <Box>
               <Box sx={{ position: 'relative' }}>
                 <Box
@@ -383,14 +126,14 @@ const SyncQr = ({
                       width: '100%',
                       borderRadius: '24px',
                     }}
-                    value={Uri}
+                    value={uri}
                     logoImage={lilo}
                     eyeColor={'#41CC5D'}
                     eyeRadius={24}
                     quietZone={20}
                   />
                 </Box>
-                {loading && (
+                {loadingString && (
                   <Box
                     sx={{
                       display: 'flex',
@@ -417,7 +160,7 @@ const SyncQr = ({
                         textAlign: 'center',
                       }}
                     >
-                      {chrome.i18n.getMessage('Scan_Successfully')}
+                      {loadingString}
                     </Typography>
                     <Typography
                       variant="body1"
@@ -430,7 +173,7 @@ const SyncQr = ({
                         textAlign: 'center',
                       }}
                     >
-                      {chrome.i18n.getMessage('Sync_in_Process')}
+                      {secondLine}
                     </Typography>
                   </Box>
                 )}
@@ -444,15 +187,12 @@ const SyncQr = ({
                   textAlign: 'center',
                 }}
               >
-                {/* {chrome.i18n.getMessage('appDescription')} {' '} */}
                 {chrome.i18n.getMessage('Scan_QR_Code_with_Mobile')}
               </Typography>
             </Box>
           )}
         </Box>
       </Box>
-
-      {/* <Box sx={{ flexGrow: 1 }} /> */}
     </>
   );
 };
